Split build command declared file read tests
